Style: 테스트 파일 및 모킹 데이터 스타일 수정
