+++ src/mocks/api/data/pageResultData.ts
@@ -1,0 +1,62 @@
+export const pageResultData = {
+  content: [
+    {
+      postId: 12,
+      title: '제목90',
+      content: '내용',
+      username: '김현율',
+      createdAt: '2023-06-28T07:24:32',
+      commentCount: 1,
+      favoriteCount: 2,
+      thumbnail:
+        'https://img4.daumcdn.net/thumb/R658x0.q70/?fname=https://t1.daumcdn.net/news/202301/26/NEWS1/20230126173028141pwuj.jpg',
+    },
+    {
+      postId: 13,
+      title: '제목12',
+      content: '내용',
+      username: '김현율',
+      createdAt: '2023-06-28T07:24:32',
+      commentCount: 1,
+      favoriteCount: 2,
+      thumbnail:
+        'https://img4.daumcdn.net/thumb/R658x0.q70/?fname=https://t1.daumcdn.net/news/202301/26/NEWS1/20230126173028141pwuj.jpg',
+    },
+    {
+      postId: 14,
+      title: '제목12',
+      content: '내용',
+      username: '김현율',
+      createdAt: '2023-06-28T07:24:32',
+      commentCount: 1,
+      favoriteCount: 2,
+      thumbnail:
+        'https://img4.daumcdn.net/thumb/R658x0.q70/?fname=https://t1.daumcdn.net/news/202301/26/NEWS1/20230126173028141pwuj.jpg',
+    },
+  ],
+  pageable: {
+    sort: {
+      empty: true,
+      unsorted: true,
+      sorted: false,
+    },
+    offset: 80,
+    pageSize: 20,
+    pageNumber: 4,
+    paged: true,
+    unpaged: false,
+  },
+  last: true,
+  totalPages: 5,
+  totalElements: 75,
+  size: 15,
+  number: 1,
+  sort: {
+    empty: true,
+    unsorted: true,
+    sorted: false,
+  },
+  first: false,
+  numberOfElements: 20,
+  empty: false,
+};
