--- src/mocks/api/data/pagetResultData.ts
@@ -1,179 +1,0 @@
-export const pageResultData = {
-  content: [
-    {
-      postId: 76,
-      title: '제목26',
-      content: '본문 내용26',
-      createdAt: '2023-06-10T20:30:02',
-      tagNames: ['소설', '영화'],
-      username: '테스트 유저',
-      likes: 4,
-      comment: 2,
-    },
-    {
-      postId: 76,
-      title: '제목26',
-      content: '본문 내용26',
-      createdAt: '2023-06-10T20:30:02',
-      tagNames: ['소설', '영화'],
-      username: '테스트 유저',
-      likes: 4,
-      comment: 2,
-    },
-    {
-      postId: 76,
-      title: '제목26',
-      content: '본문 내용26',
-      createdAt: '2023-06-10T20:30:02',
-      tagNames: ['소설', '영화'],
-      username: '테스트 유저',
-      likes: 4,
-      comment: 2,
-    },
-    {
-      postId: 76,
-      title: '제목26',
-      content: '본문 내용26',
-      createdAt: '2023-06-10T20:30:02',
-      tagNames: ['소설', '영화'],
-      username: '테스트 유저',
-      likes: 4,
-      comment: 2,
-    },
-    {
-      postId: 76,
-      title: '제목26',
-      content: '본문 내용26',
-      createdAt: '2023-06-10T20:30:02',
-      tagNames: ['소설', '영화'],
-      username: '테스트 유저',
-      likes: 4,
-      comment: 2,
-    },
-    {
-      postId: 76,
-      title: '제목26',
-      content: '본문 내용26',
-      createdAt: '2023-06-10T20:30:02',
-      tagNames: ['소설', '영화'],
-      username: '테스트 유저',
-      likes: 4,
-      comment: 2,
-    },
-    {
-      postId: 76,
-      title: '제목26',
-      content: '본문 내용26',
-      createdAt: '2023-06-10T20:30:02',
-      tagNames: ['소설', '영화'],
-      username: '테스트 유저',
-      likes: 4,
-      comment: 2,
-    },
-    {
-      postId: 76,
-      title: '제목26',
-      content: '본문 내용26',
-      createdAt: '2023-06-10T20:30:02',
-      tagNames: ['소설', '영화'],
-      username: '테스트 유저',
-      likes: 4,
-      comment: 2,
-    },
-    {
-      postId: 76,
-      title: '제목26',
-      content: '본문 내용26',
-      createdAt: '2023-06-10T20:30:02',
-      tagNames: ['소설', '영화'],
-      username: '테스트 유저',
-      likes: 4,
-      comment: 2,
-    },
-    {
-      postId: 76,
-      title: '제목26',
-      content: '본문 내용26',
-      createdAt: '2023-06-10T20:30:02',
-      tagNames: ['소설', '영화'],
-      username: '테스트 유저',
-      likes: 4,
-      comment: 2,
-    },
-    {
-      postId: 76,
-      title: '제목26',
-      content: '본문 내용26',
-      createdAt: '2023-06-10T20:30:02',
-      tagNames: ['소설', '영화'],
-      username: '테스트 유저',
-      likes: 4,
-      comment: 2,
-    },
-    {
-      postId: 76,
-      title: '제목26',
-      content: '본문 내용26',
-      createdAt: '2023-06-10T20:30:02',
-      tagNames: ['소설', '영화'],
-      username: '테스트 유저',
-      likes: 4,
-      comment: 2,
-    },
-    {
-      postId: 76,
-      title: '제목26',
-      content: '본문 내용26',
-      createdAt: '2023-06-10T20:30:02',
-      tagNames: ['소설', '영화'],
-      username: '테스트 유저',
-      likes: 4,
-      comment: 2,
-    },
-    {
-      postId: 76,
-      title: '제목26',
-      content: '본문 내용26',
-      createdAt: '2023-06-10T20:30:02',
-      tagNames: ['소설', '영화'],
-      username: '테스트 유저',
-      likes: 4,
-      comment: 2,
-    },
-    {
-      postId: 76,
-      title: '제목26',
-      content: '본문 내용26',
-      createdAt: '2023-06-10T20:30:02',
-      tagNames: ['소설', '영화'],
-      username: '테스트 유저',
-      likes: 4,
-      comment: 2,
-    },
-  ],
-  pageable: {
-    sort: {
-      empty: true,
-      unsorted: true,
-      sorted: false,
-    },
-    offset: 80,
-    pageSize: 20,
-    pageNumber: 4,
-    paged: true,
-    unpaged: false,
-  },
-  last: true,
-  totalPages: 5,
-  totalElements: 75,
-  size: 15,
-  number: 1,
-  sort: {
-    empty: true,
-    unsorted: true,
-    sorted: false,
-  },
-  first: false,
-  numberOfElements: 20,
-  empty: false,
-};
