+++ src/mocks/handlers.ts
@@ -1,4 +1,4 @@
 import authApi from './api/authApi';
-import pageApi from './api/pagetApi';
+import pageApi from './api/pageApi';
 
 export const handlers = [...authApi, ...pageApi];
